reorganizes queries
